GCC fixes

+++ cpp/src/barretenberg/proof_system/circuit_builder/goblin_translator_circuit_builder.hpp
@@ -8,8 +8,8 @@
  * @copyright Copyright (c) 2023
  *
  */
-#include "barretenberg/proof_system/arithmetization/arithmetization.hpp"
 #include "barretenberg/ecc/curves/bn254/fq.hpp"
+#include "barretenberg/proof_system/arithmetization/arithmetization.hpp"
 #include "circuit_builder_base.hpp"
 #include <array>
 namespace proof_system {
@@ -360,7 +360,7 @@ class GoblinTranslatorCircuitBuilder : CircuitBuilderBase<arithmetization::Gobli
                                                           size_t number_of_elements) {
             ASSERT(number_of_elements <= array_size);
             for (size_t i = 0; i < number_of_elements; i++) {
-                wires[starting_wire + i].push_back(add_variable(input[i]));
+                wires[static_cast<size_t>(starting_wire) + i].push_back(add_variable(input[i]));
             }
         };
         lay_limbs_in_row(acc_step.P_x_microlimbs[0], P_X_LOW_LIMBS_RANGE_CONSTRAINT_0, NUM_MICRO_LIMBS);
